Shut down gracefully & restart crashed monitors
When the process receives SIGTERM or SIGINT, it will first shut down all
containers it has spawned before shutting down itself. This is useful
for `docker-compose`, as the spawned containers are not part of the
compose project and thus not automatically shut down with
`docker-compose down`.

This "cleanup" step is no longer done on startup, as this is
unnecessary. If there are valid running monitor containers, they may
continue to run under the supervision of the new process.

Crashed and removed containers will now also be automatically restarted
while the process is running.